requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.6
dnspython==2.8.0
fastapi==0.135.1
h11==0.16.0
httptools==0.7.1
idna==3.11
motor==3.7.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
pyotp==2.9.0
python-dotenv==1.2.2
PyYAML==6.0.3
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
argon2-cffi==23.1.0
cryptography==42.0.5
